feat: Add optional autocluster with options + Add wav slice download + Add dynamic fft size + Move `h5wasm` to web worker + Fix pairings data sorting + Various attempts to improve Vue+TS QoL + Various code quality improvements

+++ processing/processing/actions/g_volumes.py
@@ -13,40 +13,41 @@ integrations = storage.get_integrations_seconds()
 volumes = storage.read_volumes()
 meta_properties = storage.read_meta_properties()
 
-storage.delete_volumes()
+if len(volumes) > 0:
+    storage.delete_volumes()
 
-print_new_line()
-print(f'Volumes loading: {[v for v in volumes]}')
+    print_new_line()
+    print(f'Volumes loading: {[v for v in volumes]}')
 
-timer = Timer(
-    len(bands) * len(integrations) * len(volumes) * len(meta_properties)
-)
+    timer = Timer(
+        len(bands) * len(integrations) * len(volumes) * len(meta_properties)
+    )
 
-for band in bands:
-    for integration in integrations:
-        grouped_features = storage.read_grouped_features_all_files(
-            band=band,
-            integration=integration,
-            unwrap=True,
-        )
+    for band in bands:
+        for integration in integrations:
+            grouped_features = storage.read_grouped_features_all_files(
+                band=band,
+                integration=integration,
+                unwrap=True,
+            )
 
-        meta_values = storage.read_meta_values(band, integration)
+            meta_values = storage.read_meta_values(band, integration)
 
-        for volume_index, volume_name in enumerate(volumes):
-            for meta_index in storage.enumerate_meta_properties():
-                meta_property = meta_properties[meta_index]
-                meta_property_values = meta_values[meta_index]
+            for volume_index, volume_name in enumerate(volumes):
+                for meta_index in storage.enumerate_meta_properties():
+                    meta_property = meta_properties[meta_index]
+                    meta_property_values = meta_values[meta_index]
 
-                v = Volume(
-                    name=volume_name,
-                    band=band,
-                    integration=integration,
-                    volume_index=volume_index,
-                    meta_index=meta_index,
-                    features=grouped_features,
-                    labels=meta_property_values,
-                )
+                    v = Volume(
+                        name=volume_name,
+                        band=band,
+                        integration=integration,
+                        volume_index=volume_index,
+                        meta_index=meta_index,
+                        features=grouped_features,
+                        labels=meta_property_values,
+                    )
 
-                v.calculate()
-                v.store(storage)
-                timer.progress()
+                    v.calculate()
+                    v.store(storage)
+                    timer.progress()
